Split marm-console's Knowledge.tsx into per-component files
Pure refactor, no behavior changes. Extracted via exact index-based
line-slicing (all 6 moved bodies verified byte-identical to the
original before any import cleanup) -- same methodology as the
Memory.tsx split.

- src/pages/Knowledge.tsx keeps only the KnowledgePage shell:
  871 -> 45 lines.
- src/components/knowledge/shared.tsx (new, 41 lines):
  DEFAULT_HIDDEN_PREDICATES, TYPE_COLORS, OTHER_TYPE_COLOR, typeColor,
  nodeRadius, mergeNeighborhoods.
- src/components/knowledge/GraphViz.tsx (new, 291 lines): the
  canvas/react-force-graph-2d visualization component alone -- the
  largest and most self-contained piece.
- src/components/knowledge/ExplorerTab.tsx (new, 340 lines):
  ProvenancePanel + ExplorerTab -- kept together since ProvenancePanel
  is only ever rendered by ExplorerTab, receiving all its data as
  props.
- src/components/knowledge/BuildAndDuplicates.tsx (new, 172 lines):
  BuildConceptsDialog + DuplicatesTab -- functionally unrelated but
  both small and both used directly by the page shell, grouped for
  file-count balance rather than functional coupling.

Split shape differs from the Memory.tsx split (flat sibling tabs) since
this file has real parent-child coupling (GraphViz/ProvenancePanel are
children of ExplorerTab, not sibling tabs) -- documented in
docs/current/knowledge-tsx-module-split.md.

Each extracted component/const referenced across file boundaries
gained `export` (mechanical TS requirement). One real gap caught by
tsc --noEmit during implementation: ExplorerTab.tsx's import list
initially omitted DEFAULT_HIDDEN_PREDICATES from shared.tsx despite
using it -- fixed before commit.

Validated: tsc --noEmit clean, npm run build succeeds, and a real
Playwright smoke pass against the running dev server confirmed both
tabs render and the Build Concepts dialog opens with a working scope
selector, zero JS console errors.
